Reorganize repeater views
A bunch of them were lost in domain.views, so I pulled those in and
split the views across two files - one for repeaters, one for repeat
records
