+++ corehq/apps/domain/views.py
@@ -6,10 +6,8 @@ import logging
 import json
 import io
 
-import pytz
 from couchdbkit import ResourceNotFound
 import dateutil
-from requests.auth import HTTPBasicAuth, HTTPDigestAuth
 from django.core.exceptions import ValidationError
 from django.core.paginator import Paginator
 from django.core.validators import validate_email
@@ -43,12 +41,10 @@ from corehq.apps.case_search.models import (
     enable_case_search,
     disable_case_search,
 )
-from corehq.apps.hqwebapp.templatetags.hq_shared_tags import toggle_js_domain_cachebuster, static
+from corehq.apps.hqwebapp.templatetags.hq_shared_tags import toggle_js_domain_cachebuster
 from corehq.apps.locations.permissions import location_safe
 from corehq.apps.locations.forms import LocationFixtureForm
 from corehq.apps.locations.models import LocationFixtureConfiguration
-from corehq.motech.repeaters.models import BASIC_AUTH, DIGEST_AUTH
-from corehq.motech.repeaters.repeater_generators import RegisterGenerator
 from corehq.const import USER_DATE_FORMAT
 from corehq.apps.accounting.async_handlers import Select2BillingInfoHandler
 from corehq.apps.accounting.invoicing import DomainWireInvoiceFactory
@@ -125,25 +121,7 @@ from corehq.apps.hqwebapp.views import BaseSectionPageView, BasePageView, CRUDPa
 from corehq.apps.domain.forms import ProjectSettingsForm
 from dimagi.utils.decorators.memoized import memoized
 from dimagi.utils.web import get_ip, json_response, get_site_domain
-from corehq.apps.users.decorators import require_can_edit_web_users, require_permission
-from corehq.motech.repeaters.forms import GenericRepeaterForm, FormRepeaterForm
-from corehq.motech.repeaters.models import Repeater, RepeatRecord
-from corehq.motech.repeaters.dbaccessors import (
-    get_paged_repeat_records,
-    get_repeat_record_count,
-    get_repeat_records_by_payload_id,
-)
-from corehq.motech.repeaters.utils import get_all_repeater_types
-from corehq.motech.repeaters.const import (
-    RECORD_FAILURE_STATE,
-    RECORD_PENDING_STATE,
-    RECORD_CANCELLED_STATE,
-    RECORD_SUCCESS_STATE,
-)
-from corehq.apps.reports.generic import GenericTabularReport
-from corehq.apps.reports.dispatcher import DomainReportDispatcher
-from corehq.apps.reports.datatables import DataTablesHeader, DataTablesColumn
-from dimagi.utils.post import simple_post
+from corehq.apps.users.decorators import require_permission
 from toggle.models import Toggle
 from corehq.apps.hqwebapp.tasks import send_html_email_async
 from corehq.apps.hqwebapp.signals import clear_login_attempts
@@ -534,110 +512,6 @@ class EditOpenClinicaSettingsView(BaseProjectSettingsView):
             else:
                 messages.error(request, _('An error occurred. Please try again.'))
         return self.get(request, *args, **kwargs)
-
-
-@require_POST
-@require_can_edit_web_users
-def cancel_repeat_record(request, domain):
-    try:
-        record = RepeatRecord.get(request.POST.get('record_id'))
-    except ResourceNotFound:
-        return HttpResponse(status=404)
-    record.cancel()
-    record.save()
-    if not record.cancelled:
-        return HttpResponse(status=400)
-    return HttpResponse('OK')
-
-
-@require_POST
-@require_can_edit_web_users
-def requeue_repeat_record(request, domain):
-    try:
-        record = RepeatRecord.get(request.POST.get('record_id'))
-    except ResourceNotFound:
-        return HttpResponse(status=404)
-    record.requeue()
-    record.save()
-    if record.cancelled:
-        return HttpResponse(status=400)
-    return HttpResponse('OK')
-
-
-@require_POST
-@require_can_edit_web_users
-def drop_repeater(request, domain, repeater_id):
-    rep = Repeater.get(repeater_id)
-    rep.retire()
-    messages.success(request, "Forwarding stopped!")
-    return HttpResponseRedirect(reverse(DomainForwardingOptionsView.urlname, args=[domain]))
-
-
-@require_POST
-@require_can_edit_web_users
-def pause_repeater(request, domain, repeater_id):
-    rep = Repeater.get(repeater_id)
-    rep.pause()
-    messages.success(request, "Forwarding paused!")
-    return HttpResponseRedirect(reverse(DomainForwardingOptionsView.urlname, args=[domain]))
-
-
-@require_POST
-@require_can_edit_web_users
-def resume_repeater(request, domain, repeater_id):
-    rep = Repeater.get(repeater_id)
-    rep.resume()
-    messages.success(request, "Forwarding resumed!")
-    return HttpResponseRedirect(reverse(DomainForwardingOptionsView.urlname, args=[domain]))
-
-
-@require_POST
-@require_can_edit_web_users
-def test_repeater(request, domain):
-    url = request.POST["url"]
-    repeater_type = request.POST['repeater_type']
-    format = request.POST.get('format', None)
-    repeater_class = get_all_repeater_types()[repeater_type]
-    auth_type = request.POST.get('auth_type')
-
-    form = GenericRepeaterForm(
-        {"url": url, "format": format},
-        domain=domain,
-        repeater_class=repeater_class
-    )
-    if form.is_valid():
-        url = form.cleaned_data["url"]
-        format = format or RegisterGenerator.default_format_by_repeater(repeater_class)
-        generator_class = RegisterGenerator.generator_class_by_repeater_format(repeater_class, format)
-        generator = generator_class(repeater_class())
-        fake_post = generator.get_test_payload(domain)
-        headers = generator.get_headers()
-
-        username = request.POST.get('username')
-        password = request.POST.get('password')
-        if auth_type == BASIC_AUTH:
-            auth = HTTPBasicAuth(username, password)
-        elif auth_type == DIGEST_AUTH:
-            auth = HTTPDigestAuth(username, password)
-        else:
-            auth = None
-
-        try:
-            resp = simple_post(fake_post, url, headers=headers, auth=auth)
-            if 200 <= resp.status_code < 300:
-                return HttpResponse(json.dumps({"success": True,
-                                                "response": resp.content,
-                                                "status": resp.status_code}))
-            else:
-                return HttpResponse(json.dumps({"success": False,
-                                                "response": resp.content,
-                                                "status": resp.status_code}))
-
-        except Exception as e:
-            errors = str(e)
-        return HttpResponse(json.dumps({"success": False, "response": errors}))
-    else:
-        return HttpResponse(json.dumps({"success": False, "response": "Please enter a valid url."}))
 
 
 def autocomplete_fields(request, field):
@@ -2322,336 +2196,6 @@ class LocationFixtureConfigView(BaseAdminProjectSettingsView):
         location_settings = LocationFixtureConfiguration.for_domain(self.domain)
         form = LocationFixtureForm(instance=location_settings)
         return {'form': form}
-
-
-class DomainForwardingRepeatRecords(GenericTabularReport):
-    name = 'Repeat Records'
-    base_template = 'domain/repeat_record_report.html'
-    section_name = 'Project Settings'
-    slug = 'repeat_record_report'
-    dispatcher = DomainReportDispatcher
-    ajax_pagination = True
-    asynchronous = False
-    sortable = False
-
-    fields = [
-        'corehq.apps.reports.filters.select.RepeaterFilter',
-        'corehq.apps.reports.filters.select.RepeatRecordStateFilter',
-        'corehq.apps.reports.filters.simple.RepeaterPayloadIdFilter',
-    ]
-
-    def _make_cancel_payload_button(self, record_id):
-        return '''
-                <a
-                    class="btn btn-default cancel-record-payload"
-                    role="button"
-                    data-record-id={}>
-                    Cancel Payload
-                </a>
-                '''.format(record_id)
-
-    def _make_requeue_payload_button(self, record_id):
-        return '''
-                <a
-                    class="btn btn-default requeue-record-payload"
-                    role="button"
-                    data-record-id={}>
-                    Requeue Payload
-                </a>
-                '''.format(record_id)
-
-    def _make_view_payload_button(self, record_id):
-        return '''
-        <a
-            class="btn btn-default"
-            role="button"
-            data-record-id={}
-            data-toggle="modal"
-            data-target="#view-record-payload-modal">
-            View Payload
-        </a>
-        '''.format(record_id)
-
-    def _make_resend_payload_button(self, record_id):
-        return '''
-        <button
-            class="btn btn-default resend-record-payload"
-            data-record-id={}>
-            Resend Payload
-        </button>
-        '''.format(record_id)
-
-    def _get_state(self, record):
-        if record.state == RECORD_SUCCESS_STATE:
-            label_cls = 'success'
-            label_text = _('Success')
-        elif record.state == RECORD_PENDING_STATE:
-            label_cls = 'warning'
-            label_text = _('Pending')
-        elif record.state == RECORD_CANCELLED_STATE:
-            label_cls = 'danger'
-            label_text = _('Cancelled')
-        elif record.state == RECORD_FAILURE_STATE:
-            label_cls = 'danger'
-            label_text = _('Failed')
-        else:
-            label_cls = ''
-            label_text = ''
-
-        return (label_cls, label_text)
-
-    def _make_state_label(self, record):
-        return '''
-        <span class="label label-{}">
-            {}
-        </span>
-        '''.format(*self._get_state(record))
-
-    @property
-    def total_records(self):
-        if self.payload_id:
-            return len(self._get_all_records_by_payload())
-        else:
-            return get_repeat_record_count(self.domain, self.repeater_id, self.state)
-
-    @property
-    def shared_pagination_GET_params(self):
-        return [
-            {'name': 'repeater', 'value': self.request.GET.get('repeater')},
-            {'name': 'record_state', 'value': self.request.GET.get('record_state')},
-            {'name': 'payload_id', 'value': self.request.GET.get('payload_id')},
-        ]
-
-    def _format_date(self, date):
-        tz_utc_aware_date = pytz.utc.localize(date)
-        return tz_utc_aware_date.astimezone(self.timezone).strftime('%b %d, %Y %H:%M:%S %Z')
-
-    @memoized
-    def _get_all_records_by_payload(self):
-        # It is assumed that there are relatively few repeat records for a given payload,
-        # so this is just filtered in memory.  If that changes, we should filter in the db.
-        return [
-            r for r in get_repeat_records_by_payload_id(self.domain, self.payload_id)
-            if (not self.repeater_id or r.repeater_id == self.repeater_id)
-            and (not self.state or r.state == self.state)
-        ]
-
-    @property
-    def payload_id(self):
-        return self.request.GET.get('payload_id', None)
-
-    @property
-    def rows(self):
-        self.repeater_id = self.request.GET.get('repeater', None)
-        self.state = self.request.GET.get('record_state', None)
-        if self.payload_id:
-            end = self.pagination.start + self.pagination.count
-            records = self._get_all_records_by_payload()[self.pagination.start:end]
-        else:
-            records = get_paged_repeat_records(
-                self.domain,
-                self.pagination.start,
-                self.pagination.count,
-                repeater_id=self.repeater_id,
-                state=self.state
-            )
-        rows = [self._make_row(record) for record in records]
-        return rows
-
-    def _payload_id_and_search_link(self, payload_id):
-        return (
-            '<a href="{url}?q={payload_id}">'
-            '<img src="{flower}" title="Search in HQ" width="14px" height="14px" />'
-            '</a> {payload_id}'
-        ).format(
-            url=reverse('global_quick_find'),
-            flower=static('prelogin/images/commcare-flower.png'),
-            payload_id=payload_id,
-        )
-
-    def _make_row(self, record):
-        row = [
-            self._make_state_label(record),
-            record.repeater.get_url(record) if record.repeater else _(u'Unable to generate url for record'),
-            self._format_date(record.last_checked) if record.last_checked else '---',
-            self._format_date(record.next_check) if record.next_check else '---',
-            render_to_string('domain/repeaters/partials/attempt_history.html', {'record': record}),
-            self._make_view_payload_button(record.get_id),
-            self._make_resend_payload_button(record.get_id),
-            self._make_requeue_payload_button(record.get_id) if record.cancelled and not record.succeeded
-            else self._make_cancel_payload_button(record.get_id) if not record.cancelled
-            and not record.succeeded
-            else None
-        ]
-
-        if toggles.SUPPORT.enabled_for_request(self.request):
-            row.insert(1, self._payload_id_and_search_link(record.payload_id))
-        return row
-
-    @property
-    def headers(self):
-        columns = [
-            DataTablesColumn(_('Status')),
-            DataTablesColumn(_('URL')),
-            DataTablesColumn(_('Last sent date')),
-            DataTablesColumn(_('Retry Date')),
-            DataTablesColumn(_('Delivery Attempts')),
-            DataTablesColumn(_('View payload')),
-            DataTablesColumn(_('Resend')),
-            DataTablesColumn(_('Cancel or Requeue payload'))
-        ]
-        if toggles.SUPPORT.enabled_for_request(self.request):
-            columns.insert(1, DataTablesColumn(_('Payload ID')))
-
-        return DataTablesHeader(*columns)
-
-
-class DomainForwardingOptionsView(BaseAdminProjectSettingsView):
-    urlname = 'domain_forwarding'
-    page_title = ugettext_lazy("Data Forwarding")
-    template_name = 'domain/admin/domain_forwarding.html'
-
-    @method_decorator(require_permission(Permissions.edit_motech))
-    def dispatch(self, request, *args, **kwargs):
-        return super(BaseProjectSettingsView, self).dispatch(request, *args, **kwargs)
-
-    @property
-    def repeaters(self):
-        return [
-            (
-                r.__name__,
-                r.by_domain(self.domain),
-                r.friendly_name,
-                r.get_custom_url(self.domain)
-            )
-            for r in get_all_repeater_types().values() if r.available_for_domain(self.domain)
-        ]
-
-    @property
-    def page_context(self):
-        return {
-            'repeaters': self.repeaters,
-            'pending_record_count': RepeatRecord.count(self.domain),
-            'gefingerpoken': self.request.couch_user.is_superuser or
-                             toggles.IS_DEVELOPER.enabled(self.request.couch_user.username)
-        }
-
-
-class BaseRepeaterView(BaseAdminProjectSettingsView):
-    page_title = ugettext_lazy("Forward Data")
-    repeater_form_class = GenericRepeaterForm
-    template_name = 'domain/admin/add_form_repeater.html'
-
-    @method_decorator(require_permission(Permissions.edit_motech))
-    def dispatch(self, request, *args, **kwargs):
-        return super(BaseRepeaterView, self).dispatch(request, *args, **kwargs)
-
-    @property
-    def page_url(self):
-        return reverse(self.urlname, args=[self.domain, self.repeater_type])
-
-    @property
-    def parent_pages(self):
-        return [{
-            'title': DomainForwardingOptionsView.page_title,
-            'url': reverse(DomainForwardingOptionsView.urlname, args=[self.domain]),
-        }]
-
-    @property
-    def repeater_type(self):
-        return self.kwargs['repeater_type']
-
-    @property
-    def page_name(self):
-        return self.repeater_class.friendly_name
-
-    @property
-    @memoized
-    def repeater_class(self):
-        try:
-            return get_all_repeater_types()[self.repeater_type]
-        except KeyError:
-            raise Http404(
-                "No such repeater {}. Valid types: {}".format(
-                    self.repeater_type, list(get_all_repeater_types())
-                )
-            )
-
-    @property
-    def add_repeater_form(self):
-        return None
-
-    @property
-    def page_context(self):
-        return {
-            'form': self.add_repeater_form,
-            'repeater_type': self.repeater_type,
-        }
-
-    def initialize_repeater(self):
-        raise NotImplementedError
-
-    def make_repeater(self):
-        repeater = self.initialize_repeater()
-        return self.set_repeater_attr(repeater, self.add_repeater_form.cleaned_data)
-
-    def set_repeater_attr(self, repeater, cleaned_data):
-        repeater.domain = self.domain
-        repeater.url = cleaned_data['url']
-        repeater.auth_type = cleaned_data['auth_type'] or None
-        repeater.username = cleaned_data['username']
-        repeater.password = cleaned_data['password']
-        repeater.format = cleaned_data['format']
-        return repeater
-
-    def post_save(self, request, repeater):
-        pass
-
-    def post(self, request, *args, **kwargs):
-        if self.add_repeater_form.is_valid():
-            repeater = self.make_repeater()
-            repeater.save()
-            return self.post_save(request, repeater)
-        return self.get(request, *args, **kwargs)
-
-
-class AddRepeaterView(BaseRepeaterView):
-    urlname = 'add_repeater'
-
-    @property
-    @memoized
-    def add_repeater_form(self):
-        if self.request.method == 'POST':
-            return self.repeater_form_class(
-                self.request.POST,
-                domain=self.domain,
-                repeater_class=self.repeater_class
-            )
-        return self.repeater_form_class(
-            domain=self.domain,
-            repeater_class=self.repeater_class
-        )
-
-    def initialize_repeater(self):
-        return self.repeater_class()
-
-    def post_save(self, request, repeater):
-        messages.success(request, _("Forwarding set up to %s" % repeater.url))
-        return HttpResponseRedirect(reverse(DomainForwardingOptionsView.urlname, args=[self.domain]))
-
-
-class AddFormRepeaterView(AddRepeaterView):
-    urlname = 'add_form_repeater'
-    repeater_form_class = FormRepeaterForm
-
-    @property
-    def page_url(self):
-        return reverse(self.urlname, args=[self.domain])
-
-    def set_repeater_attr(self, repeater, cleaned_data):
-        repeater = super(AddFormRepeaterView, self).set_repeater_attr(repeater, cleaned_data)
-        repeater.include_app_id_param = self.add_repeater_form.cleaned_data['include_app_id_param']
-        return repeater
 
 
 class BaseInternalDomainSettingsView(BaseProjectSettingsView):
