+++ corehq/motech/repeaters/views/__init__.py
@@ -1,0 +1,21 @@
+from .repeaters import (
+    AddCaseRepeaterView,
+    AddCustomSOAPCaseRepeaterView,
+    AddCustomSOAPLocationRepaterView,
+    AddFormRepeaterView,
+    AddRepeaterView,
+    DomainForwardingOptionsView,
+    EditCaseRepeaterView,
+    EditFormRepeaterView,
+    EditRepeaterView,
+    drop_repeater,
+    pause_repeater,
+    resume_repeater,
+    test_repeater,
+)
+from .repeat_records import (
+    DomainForwardingRepeatRecords,
+    RepeatRecordView,
+    cancel_repeat_record,
+    requeue_repeat_record,
+)
